requirements.txt
autopep8==1.5.5
click==7.1.2
colorama==0.4.4
copier==5.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pathspec==0.8.1
plumbum==1.7.0
pycodestyle==2.6.0
pydantic==1.8
pyparsing==2.4.7
PyYAML==5.4.1
pyyaml-include==1.2.post2
regex==2020.11.13
toml==0.10.2
typing-extensions==3.7.4.3
